BasicFlow: fix bug for PSH and URG flags counting
In the first packet of a flow which is handled in
BasicFlow.java:firstPacket, the flags PSH and URG are counted
correctly.
But in the second and after packets of a flow which is handled in
BasicFlow.java:addPacket, the flags PSH and URG are not counted
correctly.
This commit counts PSH and URG flags correctly for the second and
after packets.