BasicFlow: fix bugs for the source ip of a flow
In the first packet of a flow which is handled in
BasicFlow.java:firstPacket, when calling function:updateFlowBulk, the
BasicFlow.src is null, but the BasicFlow.src is used to compare with
the source ip of the first packet.
In function:updateFlowBulk, in order to compare with BasicFlow.src
and the source ip of the first packet, the code use operator ==
rather than Arrays.equals.
This commit fixes these 2 bugs.